Move files to TelemetryClient folder
